fix(crm): update relative imports after moving CRM files to convex/crm/ (closes #4405)
All 17 CRM files that were moved from convex/ to convex/crm/ retained their
old ./ relative imports for _generated/, _helpers/, auditLog, notifications,
and inventory. Updated them to ../ so they resolve correctly from the new
subdirectory location. Convex backend typecheck (tsc -p convex/tsconfig.json)
passes cleanly after this fix.
