+++ convex/crm/contacts.ts
@@ -1,10 +1,10 @@
-import { action, internalMutation } from "./_generated/server";
-import { internal } from "./_generated/api";
-import { createSupabaseDb } from "./_helpers/supabaseDb";
+import { action, internalMutation } from "../_generated/server";
+import { internal } from "../_generated/api";
+import { createSupabaseDb } from "../_helpers/supabaseDb";
 import { v } from "convex/values";
-import { logActivity } from "./_helpers/activities";
-import { logAudit } from "./auditLog";
-import { Id } from "./_generated/dataModel";
+import { logActivity } from "../_helpers/activities";
+import { logAudit } from "../auditLog";
+import { Id } from "../_generated/dataModel";
 
 // Dual-write refs removed — Supabase is now primary for contact writes
 // list query removed — browser reads contacts directly from Supabase via use-supabase-contacts.ts
