+++ convex/crm/emailAccounts.ts
@@ -1,6 +1,6 @@
-import { action } from "./_generated/server";
-import { internal } from "./_generated/api";
-import { createSupabaseDb } from "./_helpers/supabaseDb";
+import { action } from "../_generated/server";
+import { internal } from "../_generated/api";
+import { createSupabaseDb } from "../_helpers/supabaseDb";
 import { v } from "convex/values";
 
 // Source of truth: Supabase `email_accounts`. The Convex `emailAccounts`
