+++ convex/crm/leads.ts
@@ -1,13 +1,13 @@
-import { action, internalMutation } from "./_generated/server";
-import { internal } from "./_generated/api";
-import { createSupabaseDb } from "./_helpers/supabaseDb";
+import { action, internalMutation } from "../_generated/server";
+import { internal } from "../_generated/api";
+import { createSupabaseDb } from "../_helpers/supabaseDb";
 import { v } from "convex/values";
 import { paginationOptsValidator } from "convex/server";
-import { logActivity } from "./_helpers/activities";
+import { logActivity } from "../_helpers/activities";
 import { leadStatusValidator, leadPriorityValidator } from "@cvx/schema";
-import { logAudit } from "./auditLog";
-import { createNotificationDirect } from "./notifications";
-import { Id } from "./_generated/dataModel";
+import { logAudit } from "../auditLog";
+import { createNotificationDirect } from "../notifications";
+import { Id } from "../_generated/dataModel";
 
 // Dual-write refs removed — Supabase is now primary for lead writes
