+++ convex/crm/pipelineStageActions.ts
@@ -1,7 +1,7 @@
-import { action } from "./_generated/server";
+import { action } from "../_generated/server";
 import { v } from "convex/values";
-import { internal } from "./_generated/api";
-import { createSupabaseDb } from "./_helpers/supabaseDb";
+import { internal } from "../_generated/api";
+import { createSupabaseDb } from "../_helpers/supabaseDb";
 
 const configValidator = v.object({
   activityTypeId: v.optional(v.string()),
